+++ packages/adhere-ui-suspense/es/index.js
@@ -1,2 +1,2 @@
-import Suspense from"./suspense";export default Suspense;
+import Sync from"./suspenseSync";import Suspense from"./suspense";import ASync from"./async";Suspense.Sync=Sync,Suspense.ASync=ASync;export default Suspense;
 //# sourceMappingURL=index.js.map