+++ packages/adhere-ui-suspense/es/types.d.ts
@@ -1,10 +1,38 @@
 import React from 'react';
-export interface ISuspenseProps {
-    reset: boolean;
-}
-export interface ISuspenseState {
+export interface ISuspense {
+    fetchData?: fetchData;
+    showLoading: showLoading;
+    renderInner: renderInner;
     isFirst: boolean;
     isFirstLoading: boolean;
+}
+export interface ISuspenseProps {
+    reset: boolean;
+    firstLoading: React.ReactElement;
+}
+export interface ISuspenseState {
+}
+export interface ISuspenseSync {
+    isLoading: boolean;
+    reset: Function;
+}
+export interface ISuspenseSyncProps extends ISuspenseProps {
+    data: any;
+    isEmpty: () => boolean;
+    renderEmpty?: Function;
+    children?: any;
+}
+export interface ISuspenseSyncState extends ISuspenseState {
+    loading: boolean;
+}
+export interface ISuspenseASyncProps extends ISuspenseProps {
+    isEmpty: () => boolean;
+    renderEmpty?: Function;
+    children?: Function;
+    fetchData?: Function;
+}
+export interface ISuspenseASyncState extends ISuspenseState {
+    loading: boolean;
 }
 export interface fetchData {
     (): any;
@@ -14,9 +42,4 @@ export interface showLoading {
 }
 export interface renderInner {
     (): React.ReactElement | null;
-}
-export interface ISuspense {
-    fetchData: fetchData;
-    showLoading: showLoading;
-    renderInner: renderInner;
 }