+++ packages/adhere-ui-suspense/src/index.ts
@@ -1,3 +1,8 @@
+import Sync from './suspenseSync';
 import Suspense from './suspense';
+import ASync from './async';
+
+Suspense.Sync = Sync;
+Suspense.ASync = ASync;
 
 export default Suspense;